refactor(protocol): split crate modules
