+++ crates/protocol/src/event.rs
@@ -1,0 +1,27 @@
+use serde::Serialize;
+
+#[derive(Debug, Serialize)]
+#[serde(tag = "type", rename_all = "snake_case")]
+pub enum DaemonEvent<'message> {
+    JobStarted {
+        job_id: &'message str,
+        kind: &'message str,
+        scope: &'message str,
+    },
+    Progress {
+        job_id: &'message str,
+        message: &'message str,
+    },
+    Log {
+        job_id: &'message str,
+        message: &'message str,
+    },
+    JobCompleted {
+        job_id: &'message str,
+        summary: &'message str,
+    },
+    JobFailed {
+        job_id: &'message str,
+        error: &'message str,
+    },
+}
